# contracts/openapi/eyenet.v1.yaml
openapi: 3.1.0
info:
  title: EYENET HTTP API
  version: "1.0.0-pre"
  summary: Operator-grade evidence platform — REST + SSE surface
  description: |
    EYENET v1 HTTP API. Authoritative source of truth: `API_PLAN.md`.

    Operational invariants surfaced here:
      * All errors are `application/problem+json` (RFC 7807) using `ProblemDetail`.
      * All list endpoints are opaque-cursor paginated; no offset.
      * All writes are bus-published after a durable audit append; clients receive `202 + WriteAccepted` and must poll.
      * SSE streams declare per-event payloads via the `x-eyenet-sse-events` operation extension.
      * Bearer auth: short-lived RS256 access JWT OR a Personal Access Token (`eyenet_pat_...`).
        Authority (scopes) is resolved server-side per request — never trust scope claims in a token.
      * HTTP/2 mandatory. HTTP/1.1 is refused at the listener.
  contact:
    name: EYENET maintainers
    url: https://eyenet.local/
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://eyenet.local
    description: Operator deployment (local-bind by default; reverse-proxied for TLS)
...
security:
  - bearerAuth: []
paths:
...
  /v1/actors/{actor_id}/observations:
    get:
      tags: [actors]
      operationId: actors_observations
      summary: List observations attributable to an actor.
      parameters:
        - { $ref: '#/components/parameters/ActorIdPath' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/IncludeTotal' }
        - { $ref: '#/components/parameters/SinceTs' }
        - { $ref: '#/components/parameters/UntilTs' }
      responses:
        "200":
          description: Cursor page of observation summaries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPageObservationSummary' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "503": { $ref: '#/components/responses/Unavailable' }
  /v1/actors/{actor_id}/timeline:
    get:
      tags: [actors]
      operationId: actors_timeline
      summary: Time-ordered messages and observations for an actor.
      parameters:
        - { $ref: '#/components/parameters/ActorIdPath' }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/IncludeTotal' }
        - { $ref: '#/components/parameters/SinceTs' }
        - { $ref: '#/components/parameters/UntilTs' }
      responses:
        "200":
          description: Cursor page of timeline entries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPageTimelineEntry' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "503": { $ref: '#/components/responses/Unavailable' }
...
  /v1/graph/search:
    get:
      tags: [graph]
      operationId: graph_search
      summary: Search actors by handle/alias/identifier.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 256 }
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/IncludeTotal' }
      responses:
        "200":
          description: Cursor page of matching actor summaries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPageActorSummary' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "422": { $ref: '#/components/responses/Validation' }
        "503": { $ref: '#/components/responses/Unavailable' }

  # ────────────────────────────────────────────────────────────────────────
  # 3.3 Audit
  # ────────────────────────────────────────────────────────────────────────
...
  /v1/linkages/{linkage_id}/reject:
    post:
      tags: [linkages]
      operationId: linkages_reject
      summary: Operator rejects a linkage. Publishes `attribution.linkage.rejected`.
      parameters:
        - { $ref: '#/components/parameters/LinkageIdPath' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LinkageDecisionRequest' }
      responses:
        "202": { $ref: '#/components/responses/WriteAcceptedResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
        "422": { $ref: '#/components/responses/Validation' }
        "503": { $ref: '#/components/responses/Unavailable' }
...
  /v1/stream/linkages:
    get:
      tags: [stream]
      operationId: stream_linkages
      summary: SSE stream of linkage state-transition events.
      parameters:
        - { $ref: '#/components/parameters/LastEventId' }
        - { $ref: '#/components/parameters/StreamTokenQuery' }
      x-eyenet-sse-events:
        - subject: attribution.linkage.proposed
          payload_schema: '#/components/schemas/LinkageProposedEvent'
        - subject: attribution.linkage.confirmed
          payload_schema: '#/components/schemas/LinkageStateChangedEvent'
        - subject: attribution.linkage.rejected
          payload_schema: '#/components/schemas/LinkageStateChangedEvent'
        - subject: attribution.linkage.suspected
          payload_schema: '#/components/schemas/LinkageStateChangedEvent'
        - subject: stream.gap
          payload_schema: '#/components/schemas/StreamGapEvent'
      responses:
        "200":
          description: SSE stream (`text/event-stream`).
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /v1/stream/personas:
    get:
      tags: [stream]
      operationId: stream_personas
      summary: SSE stream of persona-update events.
      parameters:
        - { $ref: '#/components/parameters/LastEventId' }
        - { $ref: '#/components/parameters/StreamTokenQuery' }
      x-eyenet-sse-events:
        - subject: attribution.persona.updated
          payload_schema: '#/components/schemas/PersonaUpdatedEvent'
        - subject: stream.gap
          payload_schema: '#/components/schemas/StreamGapEvent'
      responses:
        "200":
          description: SSE stream (`text/event-stream`).
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
...
  /v1/stream/all:
    get:
      tags: [stream]
      operationId: stream_all
      summary: Multiplexed SSE stream — union of permitted topics.
      description: |
        Caller selects topics via repeated `?topic=` query params. The server
        intersects with the caller's granted `stream:*` scopes; topics for which
        the caller lacks scope are silently dropped (no leakage).
      parameters:
        - name: topic
          in: query
          required: true
          explode: true
          schema:
            type: array
            minItems: 1
            maxItems: 16
            items: { $ref: '#/components/schemas/StreamTopic' }
        - { $ref: '#/components/parameters/LastEventId' }
        - { $ref: '#/components/parameters/StreamTokenQuery' }
      responses:
        "200":
          description: SSE stream (`text/event-stream`).
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "422": { $ref: '#/components/responses/Validation' }

  # ────────────────────────────────────────────────────────────────────────
  # 3.6 Health / metrics
  # ────────────────────────────────────────────────────────────────────────
...
  /v1/audit/file-access/by-user:
    get:
      tags: [audit]
      operationId: audit_file_access_by_user
      summary: Per-operator file-access history over a time window.
      description: |
        The "Alice is leaving the team; what did she touch?" report (§5.8).
        Returned envelope uses the same `FileAccessExoneration` shape — the
        `content_hash` field carries a sentinel value when querying by user.
      parameters:
        - { $ref: '#/components/parameters/SystemUserIdQuery' }
        - { $ref: '#/components/parameters/SinceTs' }
        - { $ref: '#/components/parameters/UntilTs' }
      responses:
        "200":
          description: Signed per-user exoneration envelope.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileAccessExoneration' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "422": { $ref: '#/components/responses/Validation' }
        "503": { $ref: '#/components/responses/Unavailable' }
...
  /v1/clearance/grants:
    get:
      tags: [clearance]
      operationId: clearance_list_grants
      summary: List clearance grants, filterable by user and active state.
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/IncludeTotal' }
        - { $ref: '#/components/parameters/ActiveOnlyQuery' }
        - name: user_id
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: scope
          in: query
          required: false
          schema: { $ref: '#/components/schemas/ClearanceScope' }
      responses:
        "200":
          description: Page of clearance grants.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursorPageClearanceGrantSummary' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "503": { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [clearance]
      operationId: clearance_grant
      summary: Grant a clearance scope to a user. Requires `admin:clearance`.
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClearanceGrantRequest' }
      responses:
        "201":
          description: Grant created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClearanceGrantDetail' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "409": { $ref: '#/components/responses/Conflict' }
        "422": { $ref: '#/components/responses/Validation' }
        "503": { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT or eyenet_pat_<prefix>_<secret>"
      description: |
        Accepts EITHER:
          * An RS256 access JWT issued by `/v1/auth/login` (15-minute TTL).
          * A Personal Access Token of the form `eyenet_pat_<22-char-prefix>_<32-char-secret>`.

        Scopes are resolved server-side at request time from `system_user_scope`.
        Tokens never carry scope claims.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque pagination cursor returned by a prior page.
      schema: { type: string, maxLength: 4096 }
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Default 50, max 500.
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    IncludeTotal:
      name: include_total
      in: query
      required: false
      description: When `1`, include a best-effort `estimated_total` (capped at 10000).
      schema: { type: integer, enum: [0, 1] }
    SinceTs:
      name: since
      in: query
      required: false
      schema: { type: string, format: date-time }
    UntilTs:
      name: until
      in: query
      required: false
      schema: { type: string, format: date-time }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated key (uuid7 recommended). Replays return the stored response.
      schema: { type: string, minLength: 8, maxLength: 128 }
    LastEventId:
      name: Last-Event-ID
      in: header
      required: false
      description: SSE resume cursor (uuid7 of last received event).
      schema: { type: string, maxLength: 64 }
    StreamTokenQuery:
      name: stream_token
      in: query
      required: false
      description: Browser SSE fallback — short-lived bearer obtained from `/v1/auth/stream-token`.
      schema: { type: string, maxLength: 2048 }
    ActorIdPath:
      name: actor_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    LinkageIdPath:
      name: linkage_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    SystemUserIdQuery:
      name: user_id
      in: query
      required: true
      description: System user UUID. Required by per-user exoneration queries.
      schema: { type: string, format: uuid }
    ActiveOnlyQuery:
      name: active_only
      in: query
      required: false
      description: When `1`, return only grants whose ACTIVE(now) predicate holds (§4.8).
      schema: { type: integer, enum: [0, 1] }
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Forbidden:
      description: Authenticated but insufficient scope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Conflict:
      description: Idempotency-Key reused with a different body, or duplicate write.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Validation:
      description: Request failed schema validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    Unavailable:
      description: Audit chain, storage, or bus is degraded.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    WriteAcceptedResponse:
      description: Write accepted. Durable audit row persisted; bus fan-out is async.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WriteAccepted' }
  schemas:
...
    ProblemDetail:
      type: object
      additionalProperties: false
      required: [type, title, status, request_id]
      properties:
        type:        { type: string, format: uri }
        title:       { type: string, maxLength: 256 }
        status:      { type: integer, minimum: 100, maximum: 599 }
        detail:      { type: string, nullable: true, maxLength: 2048 }
        instance:    { type: string, nullable: true, maxLength: 2048 }
        request_id:  { type: string, maxLength: 64 }
        trace_id:    { type: string, nullable: true, pattern: '^[0-9a-f]{32}$' }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ValidationError' }
          default: []
    ValidationError:
      type: object
      additionalProperties: false
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - { type: string }
              - { type: integer }
        msg:  { type: string, maxLength: 512 }
        type: { type: string, maxLength: 128 }

    # ── Pagination ────────────────────────────────────────────────────────
    CursorPageBase:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: {}
        next_cursor:     { type: string, nullable: true }
        estimated_total: { type: integer, nullable: true, minimum: 0 }

    # Concrete CursorPage variants (the OpenAPI way — no generics).
...
    CursorPageObservationSummary:
      allOf:
        - { $ref: '#/components/schemas/CursorPageBase' }
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ObservationSummary' }
    CursorPageTimelineEntry:
      allOf:
        - { $ref: '#/components/schemas/CursorPageBase' }
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/TimelineEntry' }
...
    CursorPageActorSummary:
      allOf:
        - { $ref: '#/components/schemas/CursorPageBase' }
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ActorSummary' }
...
    LinkageState:
      type: string
      enum: [proposed, suspected, confirmed, rejected, superseded]
    StreamTopic:
      type: string
      enum:
        - attribution.linkage
        - attribution.persona
        - eyenet.audit
        - eyenet.control

    # ── Auth ──────────────────────────────────────────────────────────────
...
    ActorSummary:
      type: object
      additionalProperties: false
      required: [actor_id, primary_handle]
      properties:
        actor_id:       { type: string, format: uuid }
        primary_handle: { type: string, maxLength: 256 }
        platforms:
          type: array
          items: { type: string, maxLength: 32 }
        score:          { type: number, format: float, nullable: true }
...
    ObservationSummary:
      type: object
      additionalProperties: false
      required: [observation_id, kind, ts, sensitivity]
      properties:
        observation_id: { type: string, format: uuid }
        kind:           { type: string, maxLength: 64 }
        ts:             { type: string, format: date-time }
        score:          { type: number, format: float, nullable: true }
        primitive:      { type: string, maxLength: 64, nullable: true }
        sensitivity:    { $ref: '#/components/schemas/SensitivityTier' }
        attachment_blob_id:
          description: Present when this observation has a binary attachment. Use `/v1/attachments/{blob_id}/manifest` to inspect before fetching.
          type: string
          format: uuid
          nullable: true
    TimelineEntry:
      type: object
      additionalProperties: false
      required: [ts, kind, id]
      properties:
        ts:      { type: string, format: date-time }
        kind:    { type: string, enum: [message, observation] }
        id:      { type: string, format: uuid }
        summary: { type: string, maxLength: 512, nullable: true }
...
    LinkageDecisionRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 1024 }
        note:   { type: string, maxLength: 4096, nullable: true }
...
    WriteAccepted:
      type: object
      additionalProperties: false
      required: [subject, event_id, applied, poll]
      properties:
        subject:  { type: string, maxLength: 128 }
        event_id: { type: string, format: uuid }
        applied:  { type: boolean }
        poll:     { type: string, maxLength: 2048 }

    # ── Stream events ─────────────────────────────────────────────────────
    LinkageProposedEvent:
      type: object
      additionalProperties: false
      required: [event_id, linkage_id, actor_a_id, actor_b_id, score, ts]
      properties:
        event_id:   { type: string, format: uuid }
        linkage_id: { type: string, format: uuid }
        actor_a_id: { type: string, format: uuid }
        actor_b_id: { type: string, format: uuid }
        score:      { type: number, format: float }
        method:     { type: string, maxLength: 64 }
        ts:         { type: string, format: date-time }
    LinkageStateChangedEvent:
      type: object
      additionalProperties: false
      required: [event_id, linkage_id, state, ts]
      properties:
        event_id:   { type: string, format: uuid }
        linkage_id: { type: string, format: uuid }
        state:      { $ref: '#/components/schemas/LinkageState' }
        decided_by: { type: string, format: uuid, nullable: true }
        ts:         { type: string, format: date-time }
    PersonaUpdatedEvent:
      type: object
      additionalProperties: false
      required: [event_id, persona_id, ts]
      properties:
        event_id:   { type: string, format: uuid }
        persona_id: { type: string, format: uuid }
        change:     { type: string, enum: [member_added, member_removed, label_changed, created] }
        ts:         { type: string, format: date-time }
...
    StreamGapEvent:
      type: object
      additionalProperties: false
      required: [oldest_available]
      properties:
        oldest_available: { type: string, maxLength: 64 }

    # ── Health ────────────────────────────────────────────────────────────
...
    SensitivityTier:
      description: |
        Tier governs which clearance scope is required to read row content.
        Promotable but never demotable (API_PLAN §4.7).
      type: string
      enum: [normal, restricted, classified]
...
    FileServedVia:
      description: How the bytes were served — `file_access_journal.served_via` (§5.6).
      type: string
      enum: [inline_json, attachment_stream, thumbnail_only]
    FileAccessJournalEntry:
      description: Row from the `file_access_journal` (§5.6).
      type: object
      additionalProperties: false
      required:
        [access_id, audit_event_id, user_id, content_hash, content_size,
         tier, served_at, served_via, operator_signature_verified,
         signing_pubkey_fingerprint]
      properties:
        access_id:                   { type: string, format: uuid }
        audit_event_id:              { type: string, format: uuid }
        user_id:                     { type: string, format: uuid }
        grant_id:                    { type: string, format: uuid, nullable: true }
        content_hash:                { type: string, pattern: '^[0-9a-f]{64}$' }
        content_size:                { type: integer, minimum: 0 }
        content_mime:                { type: string, maxLength: 128, nullable: true }
        tier:                        { $ref: '#/components/schemas/SensitivityTier' }
        served_at:                   { type: string, format: date-time }
        served_via:                  { $ref: '#/components/schemas/FileServedVia' }
        operator_signature_verified: { type: boolean }
        signing_pubkey_fingerprint:  { type: string, pattern: '^[0-9a-f]{16}$' }
    FileAccessExoneration:
      description: |
        Signed assertion against the journal head (§5.8). An empty `accesses`
        array is a positive cryptographic proof of non-access at the time
        `query_time` for the journal at head `journal_head_at_query`.
      type: object
      additionalProperties: false
      required:
        [content_hash, query_time, journal_head_at_query, accesses, exoneration_signature]
      properties:
        content_hash:           { type: string, pattern: '^[0-9a-f]{64}$' }
        query_time:             { type: string, format: date-time }
        journal_head_at_query:  { type: string, pattern: '^[0-9a-f]{64}$' }
        accesses:
          type: array
          items: { $ref: '#/components/schemas/FileAccessJournalEntry' }
        exoneration_signature:  { type: string, pattern: '^ed25519:[A-Za-z0-9_-]{86,90}={0,2}$' }

    # ── External anchoring (API_PLAN §5.9) ────────────────────────────────
...
    ClearanceScope:
      description: |
        Grant-only scopes that flow through the §4.8 clearance-grant lifecycle:
        bounded expiry (≤90 days), mandatory reason, individually revocable.
        `read:restricted` / `read:classified` gate sensitive-evidence reads (§4.7).
        `admin:reclassify` authorises sensitivity-tier promotion (§4.9).
        `admin:case` authorises administrative case actions (archive, reopen-from-archived,
        override-collaborator) (§4.10).
      type: string
      enum: [read:restricted, read:classified, admin:reclassify, admin:case]
    ClearanceGrantRequest:
      description: |
        Body for `POST /v1/clearance/grants`. Caller must hold `admin:clearance`.
        `expires_at` is mandatory and capped at `granted_at + 90 days` server-side
        (§4.8). `reason` length ≥ 16 — "ok" is not a reason. `case_refs` (optional)
        anchors the grant to one or more §4.10 cases; server validates each exists
        and is `open`.
      type: object
      additionalProperties: false
      required: [user_id, scope, reason, expires_at]
      properties:
        user_id:    { type: string, format: uuid }
        scope:      { $ref: '#/components/schemas/ClearanceScope' }
        reason:     { type: string, minLength: 16, maxLength: 1024 }
        expires_at: { type: string, format: date-time }
        case_refs:
          type: array
          maxItems: 32
          items: { type: string, format: uuid }
          description: Cases this grant authorises work under (§4.10.7).
...
    ClearanceGrantSummary:
      type: object
      additionalProperties: false
      required:
        [grant_id, user_id, scope, granted_by_user_id, granted_at, expires_at, active]
      properties:
        grant_id:           { type: string, format: uuid }
        user_id:            { type: string, format: uuid }
        scope:              { $ref: '#/components/schemas/ClearanceScope' }
        granted_by_user_id: { type: string, format: uuid }
        granted_at:         { type: string, format: date-time }
        expires_at:         { type: string, format: date-time }
        revoked_at:         { type: string, format: date-time, nullable: true }
        active:             { type: boolean, description: "Server-evaluated ACTIVE(now) predicate." }
    ClearanceGrantDetail:
      allOf:
        - { $ref: '#/components/schemas/ClearanceGrantSummary' }
        - type: object
          additionalProperties: false
          required: [reason]
          properties:
            reason:             { type: string, minLength: 16, maxLength: 1024 }
            revoked_by_user_id: { type: string, format: uuid, nullable: true }
            revocation_reason:  { type: string, maxLength: 1024, nullable: true }
            parent_grant_id:    { type: string, format: uuid, nullable: true }
    CursorPageClearanceGrantSummary:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ClearanceGrantSummary' }
        next_cursor:     { type: string, nullable: true }
        estimated_total: { type: integer, nullable: true, minimum: 0 }

    # ── Reclassification (API_PLAN §4.9) ──────────────────────────────────
...